+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Forge
+
+- Bug with using cheats in library calls
+
 ## [0.58.0] - 2026-03-18
 
 ### Forge